# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Document {
  id: number;
  name: string;
  content: string;
  createdAt: string;
}

export interface UploadDocumentBody {
  name: string;
  content: string;
}

export interface AskDocumentBody {
  question: string;
}

export interface ChatMessage {
  id: number;
  documentId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface ApiError {
  error: string;
}