# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: documents
    description: Company document management
  - name: chat
    description: AI-powered Q&A chat
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List all uploaded documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a company document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UploadDocumentBody"
      responses:
        "201":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      summary: Get a document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /documents/{id}/messages:
    get:
      operationId: listDocumentMessages
      tags: [chat]
      summary: Get Q&A message history for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /documents/{id}/ask:
    post:
      operationId: askDocument
      tags: [chat]
      summary: Ask a question about the document (SSE streaming)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AskDocumentBody"
      responses:
        "200":
          description: SSE stream of answer chunks
          content:
            text/event-stream: {}
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /documents/{id}/messages/clear:
    delete:
      operationId: clearDocumentMessages
      tags: [chat]
      summary: Clear all chat history for a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cleared
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Document:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - content
        - createdAt
    UploadDocumentBody:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
      required:
        - name
        - content
    AskDocumentBody:
      type: object
      properties:
        question:
          type: string
      required:
        - question
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - documentId
        - role
        - content
        - createdAt
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error