# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all uploaded documents
 */
export const ListDocumentsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListDocumentsResponse = zod.array(ListDocumentsResponseItem);

/**
 * @summary Upload a company document
 */
export const UploadDocumentBody = zod.object({
  name: zod.string(),
  content: zod.string(),
});

/**
 * @summary Get a document by ID
 */
export const GetDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const GetDocumentResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Delete a document
 */
export const DeleteDocumentParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get Q&A message history for a document
 */
export const ListDocumentMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListDocumentMessagesResponseItem = zod.object({
  id: zod.number(),
  documentId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.coerce.date(),
});
export const ListDocumentMessagesResponse = zod.array(
  ListDocumentMessagesResponseItem,
);

/**
 * @summary Ask a question about the document (SSE streaming)
 */
export const AskDocumentParams = zod.object({
  id: zod.coerce.number(),
});

export const AskDocumentBody = zod.object({
  question: zod.string(),
});

/**
 * @summary Clear all chat history for a document
 */
export const ClearDocumentMessagesParams = zod.object({
  id: zod.coerce.number(),
});